🔨 refactor: merge documents/transactions and ducks/transactions/helpers
