+++ src/ducks/transactions/TransactionsPage.jsx
@@ -29,13 +29,13 @@ import SelectDates, { ConnectedSelectDates } from 'components/SelectDates'
 import { fetchTransactions } from 'actions/transactions'
 import { getAppUrlById, fetchApps } from 'ducks/apps'
 import { getCategoryIdFromName } from 'ducks/categories/categoriesMap'
+import { hydrateTransaction, getDate } from 'ducks/transactions/helpers'
 import { getCategoryId } from 'ducks/categories/helpers'
 
 import Loading from 'components/Loading'
 import { Breadcrumb } from 'components/Breadcrumb'
 import BackButton from 'components/BackButton'
 
-import { hydrateTransaction } from 'documents/transaction'
 import { TransactionTableHead, TransactionsWithSelection } from './Transactions'
 import styles from './TransactionsPage.styl'
 import { TRIGGER_DOCTYPE, ACCOUNT_DOCTYPE } from 'doctypes'
